qrcode_scanner/requirement.txt
Pillow (7.2.0)
pyzbar (0.1.8)
